requirements.txt
# Skin Disease Detection - Dependencies
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data & preprocessing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Visualization & evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
